Make a sed to make up a blacklist conf for nginx
I'll eventually move this to using iptables instead on my server, but for now this will work just fine until I do.

also not sure what the 858.5.5.5 address is (it seems to not be a real address at all), but I needed to remove it otherwise nginx throws an error.